gpio: move DummyDevice and rename to MockGpioDevice
We are about to use it to simulate a device for the real binary so
lets move it out of the gpio module into its own. We also migrate the
VirtIO constants to keep the imports somewhat sane.
